Modularize the implementation
